pyproject.toml
[project]
name = "cv-agentic"
version = "0.1.0"
description = "LLM-only agentic CV tailor"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv"
]


[tool.uv]
dev-dependencies = []

